# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "연말정산 시뮬레이터 API",
    "description": "근로소득 연말정산 예상 결정세액·환급액 계산 (참고용 시뮬레이션)",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/businesses/allocate": {
      "post": {
        "tags": [
          "businesses"
        ],
        "summary": "사업장 소득금액을 손익분배비율대로 배분 (저장하지 않음)",
        "operationId": "allocate_business_api_v1_businesses_allocate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessRecord-Input"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessAllocation"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "리소스 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "입력 검증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/businesses/partnership": {
      "post": {
        "tags": [
          "businesses"
        ],
        "summary": "공동사업자 일괄 종합소득세 계산 (각자의 연말정산 결과 + 지분만큼의 사업소득)",
        "operationId": "calculate_partnership_api_v1_businesses_partnership_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessRecord-Input"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartnershipResult"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "리소스 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "입력 검증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/businesses": {
      "post": {
        "tags": [
          "businesses"
        ],
        "summary": "Create Business",
        "operationId": "create_business_api_v1_businesses_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessSave"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessRead"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "리소스 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "입력 검증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "businesses"
        ],
        "summary": "List Businesses",
        "operationId": "list_businesses_api_v1_businesses_get",
        "parameters": [
          {
            "name": "tax_year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tax Year"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BusinessSummary"
                  },
                  "title": "Response List Businesses Api V1 Businesses Get"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "리소스 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "입력 검증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/businesses/{business_id}": {
      "get": {
        "tags": [
          "businesses"
        ],
        "summary": "Get Business",
        "operationId": "get_business_api_v1_businesses__business_id__get",
        "parameters": [
          {
            "name": "business_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Business Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessRead"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "리소스 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "입력 검증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "businesses"
        ],
        "summary": "Update Business",
        "operationId": "update_business_api_v1_businesses__business_id__put",
        "parameters": [
          {
            "name": "business_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Business Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessSave"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessRead"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "리소스 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "입력 검증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "businesses"
        ],
        "summary": "Delete Business",
        "operationId": "delete_business_api_v1_businesses__business_id__delete",
        "parameters": [
          {
            "name": "business_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Business Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "리소스 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "입력 검증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BusinessAllocation": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "tax_year": {
            "type": "integer",
            "title": "Tax Year"
          },
          "revenue": {
            "type": "integer",
            "title": "Revenue"
          },
          "expenses": {
            "type": "integer",
            "title": "Expenses"
          },
          "income_amount": {
            "type": "integer",
            "title": "Income Amount"
          },
          "withholding_tax": {
            "type": "integer",
            "title": "Withholding Tax"
          },
          "expense_note": {
            "type": "string",
            "title": "Expense Note"
          },
          "partners": {
            "items": {
              "$ref": "#/components/schemas/PartnerShare"
            },
            "type": "array",
            "title": "Partners"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "tax_year",
          "revenue",
          "expenses",
          "income_amount",
          "withholding_tax",
          "expense_note",
          "partners"
        ],
        "title": "BusinessAllocation"
      },
...
      "BusinessIncome-Output": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 50,
            "title": "Name",
            "default": ""
          },
          "revenue": {
            "type": "integer",
            "maximum": 10000000000000.0,
            "minimum": 0.0,
            "title": "Revenue",
            "description": "총수입금액"
          },
          "expense_method": {
            "$ref": "#/components/schemas/ExpenseMethod",
            "default": "rate"
          },
          "expenses": {
            "type": "integer",
            "maximum": 10000000000000.0,
            "minimum": 0.0,
            "title": "Expenses",
            "description": "장부 방식의 필요경비",
            "default": 0
          },
          "expense_rate": {
            "type": "string",
            "title": "Expense Rate",
            "description": "경비율(%) — 국세청 업종별 경비율 조회",
            "default": "0"
          },
          "withholding_tax": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10000000000000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Withholding Tax",
            "description": "원천징수된 소득세. 비우면 총수입금액 × 3%로 추정"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "revenue"
        ],
        "title": "BusinessIncome"
      },
      "BusinessRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "tax_year": {
            "type": "integer",
            "title": "Tax Year"
          },
          "partner_count": {
            "type": "integer",
            "title": "Partner Count"
          },
          "income_amount": {
            "type": "integer",
            "title": "Income Amount"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "record": {
            "$ref": "#/components/schemas/BusinessRecord-Output"
          },
          "allocation": {
            "$ref": "#/components/schemas/BusinessAllocation"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "tax_year",
          "partner_count",
          "income_amount",
          "updated_at",
          "record",
          "allocation"
        ],
        "title": "BusinessRead"
      },
      "BusinessRecord-Input": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Name"
          },
          "tax_year": {
            "type": "integer",
            "maximum": 2100.0,
            "minimum": 2000.0,
            "title": "Tax Year"
          },
          "revenue": {
            "type": "integer",
            "maximum": 10000000000000.0,
            "minimum": 0.0,
            "title": "Revenue"
          },
          "expense_method": {
            "$ref": "#/components/schemas/ExpenseMethod",
            "default": "book"
          },
          "expenses": {
            "type": "integer",
            "maximum": 10000000000000.0,
            "minimum": 0.0,
            "title": "Expenses",
            "default": 0
          },
          "expense_rate": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "string"
              }
            ],
            "title": "Expense Rate",
            "default": "0"
          },
          "withholding_tax": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10000000000000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Withholding Tax",
            "description": "사업장 전체 원천징수세액. 비우면 총수입금액 × 3%로 추정"
          },
          "partners": {
            "items": {
              "$ref": "#/components/schemas/Partner"
            },
            "type": "array",
            "maxItems": 10,
            "minItems": 1,
            "title": "Partners"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "tax_year",
          "revenue",
          "partners"
        ],
        "title": "BusinessRecord",
        "description": "사업자 소득관리 레코드 (사업장 단위)."
      },
      "BusinessRecord-Output": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Name"
          },
          "tax_year": {
            "type": "integer",
            "maximum": 2100.0,
            "minimum": 2000.0,
            "title": "Tax Year"
          },
          "revenue": {
            "type": "integer",
            "maximum": 10000000000000.0,
            "minimum": 0.0,
            "title": "Revenue"
          },
          "expense_method": {
            "$ref": "#/components/schemas/ExpenseMethod",
            "default": "book"
          },
          "expenses": {
            "type": "integer",
            "maximum": 10000000000000.0,
            "minimum": 0.0,
            "title": "Expenses",
            "default": 0
          },
          "expense_rate": {
            "type": "string",
            "title": "Expense Rate",
            "default": "0"
          },
          "withholding_tax": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10000000000000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Withholding Tax",
            "description": "사업장 전체 원천징수세액. 비우면 총수입금액 × 3%로 추정"
          },
          "partners": {
            "items": {
              "$ref": "#/components/schemas/Partner"
            },
            "type": "array",
            "maxItems": 10,
            "minItems": 1,
            "title": "Partners"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "tax_year",
          "revenue",
          "partners"
        ],
        "title": "BusinessRecord",
        "description": "사업자 소득관리 레코드 (사업장 단위)."
      },
      "BusinessSave": {
        "properties": {
          "record": {
            "$ref": "#/components/schemas/BusinessRecord-Input"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "record"
        ],
        "title": "BusinessSave"
      },
      "BusinessSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "tax_year": {
            "type": "integer",
            "title": "Tax Year"
          },
          "partner_count": {
            "type": "integer",
            "title": "Partner Count"
          },
          "income_amount": {
            "type": "integer",
            "title": "Income Amount"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "tax_year",
          "partner_count",
          "income_amount",
          "updated_at"
        ],
        "title": "BusinessSummary"
      },
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse"
      },
      "ExpenseMethod": {
        "type": "string",
        "enum": [
          "book",
          "rate"
        ],
        "title": "ExpenseMethod"
      },
...
      "GlobalIncomeInput-Output": {
        "properties": {
          "tax_year": {
            "type": "integer",
            "maximum": 2100.0,
            "minimum": 2000.0,
            "title": "Tax Year"
          },
          "base": {
            "$ref": "#/components/schemas/SimulationInput",
            "description": "근로소득·인적공제·공제 입력 (근로소득이 없으면 급여 0)"
          },
          "earned_prepaid_tax": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10000000000000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Earned Prepaid Tax",
            "description": "근로소득 기납부세액(연말정산 결정세액). 비우면 연말정산 결정세액으로 계산"
          },
          "business_incomes": {
            "items": {
              "$ref": "#/components/schemas/BusinessIncome-Output"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Business Incomes",
            "default": []
          },
          "other_incomes": {
            "items": {
              "$ref": "#/components/schemas/OtherIncome"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Other Incomes",
            "default": []
          },
          "other_income_taxation": {
            "$ref": "#/components/schemas/OtherIncomeTaxation",
            "default": "auto"
          },
          "interim_prepayment": {
            "type": "integer",
            "maximum": 10000000000000.0,
            "minimum": 0.0,
            "title": "Interim Prepayment",
            "description": "중간예납세액",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "tax_year",
          "base"
        ],
        "title": "GlobalIncomeInput"
      },
      "GlobalIncomeResult": {
        "properties": {
          "tax_year": {
            "type": "integer",
            "title": "Tax Year"
          },
          "rules_verified": {
            "type": "boolean",
            "title": "Rules Verified"
          },
          "income_lines": {
            "items": {
              "$ref": "#/components/schemas/IncomeLine"
            },
            "type": "array",
            "title": "Income Lines"
          },
          "earned_income_amount": {
            "type": "integer",
            "title": "Earned Income Amount"
          },
          "business_income_amount": {
            "type": "integer",
            "title": "Business Income Amount"
          },
          "other_income_amount": {
            "type": "integer",
            "title": "Other Income Amount"
          },
          "comprehensive_income_amount": {
            "type": "integer",
            "title": "Comprehensive Income Amount"
          },
          "other_income_taxation": {
            "type": "string",
            "enum": [
              "comprehensive",
              "separate",
              "none"
            ],
            "title": "Other Income Taxation"
          },
          "separate_tax": {
            "type": "integer",
            "title": "Separate Tax"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/TaxationOption"
            },
            "type": "array",
            "title": "Options"
          },
          "prepaid": {
            "$ref": "#/components/schemas/PrepaidBreakdown"
          },
          "tax_result": {
            "$ref": "#/components/schemas/TaxResult"
          },
          "total_balance_due": {
            "type": "integer",
            "title": "Total Balance Due"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/CalcWarning"
            },
            "type": "array",
            "title": "Warnings"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "tax_year",
          "rules_verified",
          "income_lines",
          "earned_income_amount",
          "business_income_amount",
          "other_income_amount",
          "comprehensive_income_amount",
          "other_income_taxation",
          "separate_tax",
          "options",
          "prepaid",
          "tax_result",
          "total_balance_due",
          "warnings",
          "disclaimer"
        ],
        "title": "GlobalIncomeResult"
      },
...
      "OtherIncomeTaxation": {
        "type": "string",
        "enum": [
          "auto",
          "comprehensive",
          "separate"
        ],
        "title": "OtherIncomeTaxation"
      },
      "Partner": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Name"
          },
          "share": {
            "type": "integer",
            "maximum": 1000000.0,
            "minimum": 1.0,
            "title": "Share",
            "description": "손익분배비율 지분 (예: 6:4면 6과 4)"
          },
          "simulation_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Simulation Id",
            "description": "이 사업자의 연말정산 시뮬레이션 (근로소득)"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "share"
        ],
        "title": "Partner"
      },
      "PartnerResult": {
        "properties": {
          "partner": {
            "$ref": "#/components/schemas/PartnerShare"
          },
          "simulation_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Simulation Name"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          },
          "input": {
            "$ref": "#/components/schemas/GlobalIncomeInput-Output"
          },
          "result": {
            "$ref": "#/components/schemas/GlobalIncomeResult"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "partner",
          "simulation_name",
          "notes",
          "input",
          "result"
        ],
        "title": "PartnerResult"
      },
      "PartnerShare": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "share": {
            "type": "integer",
            "title": "Share"
          },
          "ratio": {
            "type": "string",
            "title": "Ratio"
          },
          "simulation_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Simulation Id"
          },
          "revenue": {
            "type": "integer",
            "title": "Revenue"
          },
          "expenses": {
            "type": "integer",
            "title": "Expenses"
          },
          "income_amount": {
            "type": "integer",
            "title": "Income Amount"
          },
          "withholding_tax": {
            "type": "integer",
            "title": "Withholding Tax"
          },
          "business_income": {
            "$ref": "#/components/schemas/BusinessIncome-Output"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "share",
          "ratio",
          "simulation_id",
          "revenue",
          "expenses",
          "income_amount",
          "withholding_tax",
          "business_income"
        ],
        "title": "PartnerShare"
      },
      "PartnershipResult": {
        "properties": {
          "allocation": {
            "$ref": "#/components/schemas/BusinessAllocation"
          },
          "partners": {
            "items": {
              "$ref": "#/components/schemas/PartnerResult"
            },
            "type": "array",
            "title": "Partners"
          },
          "combined_total_balance_due": {
            "type": "integer",
            "title": "Combined Total Balance Due"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "allocation",
          "partners",
          "combined_total_balance_due"
        ],
        "title": "PartnershipResult"
      },
...
      "SimulationInput": {
        "properties": {
          "tax_year": {
            "type": "integer",
            "maximum": 2100.0,
            "minimum": 2000.0,
            "title": "Tax Year"
          },
          "income": {
            "$ref": "#/components/schemas/Income"
          },
          "prepaid_tax": {
            "$ref": "#/components/schemas/PrepaidTax",
            "default": {
              "withholding": 0,
              "previous_employer": 0
            }
          },
          "taxpayer": {
            "$ref": "#/components/schemas/Taxpayer"
          },
          "dependents": {
            "items": {
              "$ref": "#/components/schemas/Dependent"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Dependents",
            "default": []
          },
          "deductions": {
            "$ref": "#/components/schemas/IncomeDeductionInput",
            "default": {
              "national_pension": 0,
              "health_insurance": 0,
              "employment_insurance": 0,
              "housing_rent_loan_repayment": 0,
              "long_term_mortgage_interest": 0,
              "housing_subscription": 0,
              "card": {
                "credit": 0,
                "culture": 0,
                "debit_cash": 0,
                "public_transport": 0,
                "sports_facility": 0,
                "traditional_market": 0
              },
              "venture_direct": 0,
              "venture_fund": 0,
              "national_growth_fund": 0
            }
          },
          "credits": {
            "$ref": "#/components/schemas/TaxCreditInput",
            "default": {
              "pension_savings": 0,
              "irp": 0,
              "insurance_general": 0,
              "insurance_disabled": 0,
              "medical": {
                "general": 0,
                "infertility": 0,
                "premature": 0,
                "specific": 0
              },
              "education": [],
              "donations": {
                "general": 0,
                "hometown": 0,
                "hometown_disaster": 0,
                "political": 0,
                "religious": 0,
                "special": 0
              },
              "monthly_rent": 0
            }
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "tax_year",
          "income",
          "taxpayer"
        ],
        "title": "SimulationInput"
      },
...
    }
  }
}